+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Yii 2 HTML to PDF conversion extension Change Log
 -----------------------
 
 - Bug #12: Fixed inability to setup 'tempDir' and 'fontDir' for `Mpdf` converter (klimov-paul)
-
+- Enh #15: Added method `TempFile::getContent()` (mludvik)
 
 1.0.3, April 9, 2018
 --------------------